refactor(server): #125 — extract upgrade handlers + release machinery into routes/upgrade.rs
